doWrite 작성자 정보 저장, doModify, doDelete 권한체크

+++ src/main/java/com/KoreaIT/demo/controller/UsrArticleController.java
@@ -11,6 +11,8 @@ import com.KoreaIT.demo.util.Util;
 import com.KoreaIT.demo.vo.Article;
 import com.KoreaIT.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 
@@ -23,8 +25,12 @@ public class UsrArticleController {
 	// write, 작성
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession session, String title, String body) {
 
+		if(session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요.");
+		}
+		
 		if (Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요.");
 		}
@@ -32,8 +38,8 @@ public class UsrArticleController {
 		if (Util.empty(body)) {
 			return ResultData.from("F-2", "내용을 입력해주세요.");
 		}
-
-		articleService.writeArticle(title, body);
+		
+		articleService.writeArticle((int) session.getAttribute("loginedMemberId"), title, body);
 
 		int id = articleService.getLastInsertId();
 
@@ -73,12 +79,21 @@ public class UsrArticleController {
 	// modify, 수정
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Article> doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession session, int id, String title, String body) {
 
+		if(session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		// 권한 체크
+		if(article.getMemberId() != (int) session.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-A", Util.f("%d번 게시물에 대한 권한이 없습니다.", id));
 		}
 
 		articleService.modifyArticle(id, title, body);
@@ -89,12 +104,21 @@ public class UsrArticleController {
 	// delete, 삭제
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
 
+		if(session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		// 권한 체크
+		if(article.getMemberId() != (int) session.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-A", Util.f("%d번 게시물에 대한 권한이 없습니다.", id));
 		}
 
 		articleService.deleteArticleById(id);
